+++ py/desimodel/io.py
@@ -12,8 +12,8 @@ import yaml
 import numpy as np
 import warnings
 
-from desiutil.log import get_logger, DEBUG
-log = get_logger(DEBUG)
+from desiutil.log import get_logger
+log = get_logger()
 
 _thru = dict()
 def load_throughput(channel):
@@ -63,9 +63,19 @@ def load_desiparams():
         with open(desiparamsfile) as par:
             _params = yaml.load(par)
 
-    #- for temporary backwards compability after 'exptime' -> 'exptime_dark'
-    if ('exptime' not in _params) and ('exptime_dark' in _params):
-        _params['exptime'] = _params['exptime_dark']
+        #- for temporary backwards compability after 'exptime' -> 'exptime_dark'
+        if ('exptime' not in _params) and ('exptime_dark' in _params):
+            _params['exptime'] = _params['exptime_dark']
+
+        #- Augment params with wavelength coverage from specpsf files
+        #- wavemin/max = min/max wavelength covered by *any* fiber on the CCD
+        #- wavemin/max_all = min/max wavelength covered by *all* fibers
+        for channel in ['b', 'r', 'z']:
+            hdr = fits.getheader(findfile('specpsf/psf-{}.fits'.format(channel)), 0)
+            _params['ccd'][channel]['wavemin'] = hdr['WAVEMIN']
+            _params['ccd'][channel]['wavemax'] = hdr['WAVEMAX']
+            _params['ccd'][channel]['wavemin_all'] = hdr['WMIN_ALL']
+            _params['ccd'][channel]['wavemax_all'] = hdr['WMAX_ALL']
 
     return _params
 #
@@ -121,8 +131,8 @@ def load_fiberpos():
 #
 #
 #
-_tiles = None
-def load_tiles(onlydesi=True, extra=False):
+_tiles = dict()
+def load_tiles(onlydesi=True, extra=False, tilesfile=None, cache=True):
     """Return DESI tiles structure from desimodel/data/footprint/desi-tiles.fits.
 
     Parameters
@@ -131,37 +141,59 @@ def load_tiles(onlydesi=True, extra=False):
         If ``True``, trim to just the tiles in the DESI footprint.
     extra : :class:`bool`, (default False)
         If ``True``, include extra layers with PROGRAM='EXTRA'.
+    tilesfile : (str)
+        Name of tiles file to load; or None for default.
+        Without path, look in $DESIMODEL/data/footprint, otherwise load file.
+    cache : :class:`bool`, (default True)
+        Use cache of tiles data.
     """
     global _tiles
-    if _tiles is None:
-        footprint = os.path.join(os.environ['DESIMODEL'],'data','footprint','desi-tiles.fits')
-        with fits.open(footprint) as hdulist:
-            _tiles = hdulist[1].data
+
+    if tilesfile is None:
+        tilesfile = 'desi-tiles.fits'
+
+    #- Check if tilesfile includes a path (absolute or relative)
+    tilespath, filename = os.path.split(tilesfile)
+    if tilespath == '':
+        tilesfile = os.path.join(os.environ['DESIMODEL'],'data','footprint',filename)
+
+    #- standarize path location
+    tilesfile = os.path.abspath(tilesfile)
+
+    if cache and tilesfile in _tiles:
+        tiledata = _tiles[tilesfile]
+    else:
+        with fits.open(tilesfile, memmap=False) as hdulist:
+            tiledata = hdulist[1].data
         #
         # Temporary workaround for problem identified in
         # https://github.com/desihub/desimodel/issues/30
         #
-        if any([c.bzero is not None for c in _tiles.columns]):
-            foo = [_tiles[k].dtype for k in _tiles.dtype.names]
+        if any([c.bzero is not None for c in tiledata.columns]):
+            foo = [_tiles[k].dtype for k in tiledata.dtype.names]
 
         #- Check for out-of-date tiles file
-        if np.issubdtype(_tiles['OBSCONDITIONS'].dtype, 'u2'):
+        if np.issubdtype(tiledata['OBSCONDITIONS'].dtype, 'u2'):
             import warnings
             warnings.warn('old desi-tiles.fits with uint16 OBSCONDITIONS; please update your $DESIMODEL checkout', DeprecationWarning)
 
+        #- load cache for next time
+        if cache:
+            _tiles[tilesfile] = tiledata
+
     #- Filter to only the DESI footprint if requested
-    subset = np.ones(len(_tiles), dtype=bool)
+    subset = np.ones(len(tiledata), dtype=bool)
     if onlydesi:
-        subset &= _tiles['IN_DESI'] > 0
+        subset &= tiledata['IN_DESI'] > 0
 
     #- Filter out PROGRAM=EXTRA tiles if requested
     if not extra:
-        subset &= ~np.char.startswith(_tiles['PROGRAM'], 'EXTRA')
+        subset &= ~np.char.startswith(tiledata['PROGRAM'], 'EXTRA')
 
     if np.all(subset):
-        return _tiles
+        return tiledata
     else:
-        return _tiles[subset]
+        return tiledata[subset]
 
 _platescale = None
 def load_platescale():
@@ -186,6 +218,17 @@ def load_platescale():
     ]
     _platescale = np.loadtxt(infile, usecols=[0,1,6,7], dtype=columns)
     return _platescale
+
+def reset_cache():
+    '''Reset I/O cache'''
+    global _thru, _psf, _params, _gfa, _fiberpos, _tiles, _platescale
+    _thru = dict()
+    _psf = dict()
+    _params = None
+    _gfa = None
+    _fiberpos = None
+    _tiles = dict()
+    _platescale = None
 
 def load_target_info():
     '''