+++ py/desimodel/test/test_focalplane.py
@@ -4,9 +4,10 @@
 """
 import unittest
 import numpy as np
-from ..focalplane import FocalPlane, generate_random_centroid_offsets, xy2radec, get_radius_mm, get_radius_deg
-from ..focalplane import fiber_area_arcsec2
-
+from ..focalplane import FocalPlane, generate_random_centroid_offsets, xy2radec, get_radius_mm, get_radius_deg, radec2xy
+from ..focalplane import fiber_area_arcsec2, on_gfa, on_tile_gfa, get_gfa_targets
+from .. import io
+from astropy.table import Table
 
 class TestFocalplane(unittest.TestCase):
     """Test desimodel.focalplane.
@@ -54,7 +55,7 @@ class TestFocalplane(unittest.TestCase):
         self.assertAlmostEqual(trueradius, radius, 5)
         self.assertAlmostEqual(all(trueradius1), all(radius1), 5)
     
-    def new_test_xy2radec(self):
+    def test_xy2radec_new(self):
         """Tests the consistency between the conversion functions
         radec2xy and xy2radec. Also tests the accuracy of the xy2radec
         on particular cases.
@@ -62,11 +63,11 @@ class TestFocalplane(unittest.TestCase):
         truera = 8.927313423598427
         truedec = -9.324956250231294
         newra, newdec = xy2radec(8.37, -10.65, -138.345, -333.179)
-        self.assertEqual(truera, newra, 5)
-        self.assertEqual(truedec, newdec, 5)
+        self.assertAlmostEqual(truera, newra, 5)
+        self.assertAlmostEqual(truedec, newdec, 5)
         x, y = radec2xy(8.37, -10.65, newra, newdec)
-        self.assertEqual(x, -138.345, 5)
-        self.assertEqual(y, -333.179, 5)
+        self.assertAlmostEqual(x, -138.345, 5)
+        self.assertAlmostEqual(y, -333.179, 5)
 
     def test_xy2radec(self):
         """Test the consistency between the conversion functions
@@ -102,6 +103,45 @@ class TestFocalplane(unittest.TestCase):
         y = [-0., -50., -100., -200]
         area = fiber_area_arcsec2(x, y)
         self.assertFalse(np.any(np.fabs(area-np.array([1.97314482,  1.96207294,  1.92970506,  1.81091119])) >1E-6))
+
+    def test_on_gfa(self):
+        """Tests if on_gfa returns two lists as it is supposed to"""
+        import numpy as np
+        targetindices, gfalist = on_gfa(0, 0, np.array([1.0, 0, 0, 1.0]), np.array([0, 1.0, 1.5, .5]))
+        self.assertEqual(0, len(targetindices))
+        self.assertEqual(0, len(gfalist))
+        self.assertEqual(len(targetindices), len(gfalist))
+
+    def test_on_tile_gfa(self):
+        """Tests if on_tile_gfa returns two lists as it is supposed to"""
+        #- Generate some targets near a tile; some should land on the GFAs
+        tile = io.load_tiles()[0]
+        targets = Table()
+        ntargets = 1000
+        ra = np.random.uniform(tile['RA']-1.61, tile['RA']+1.61, size=ntargets)
+        dec = np.random.uniform(tile['DEC']-1.61, tile['DEC']+1.61, size=ntargets)
+        targets['RA'] = (ra+360) % 360
+        targets['DEC'] = dec
+        targetindices, gfaid = on_tile_gfa(tile['TILEID'], targets, 120)
+        self.assertEqual(targetindices.size, gfaid.size)
+        self.assertGreater(targetindices.size, 0)
+        self.assertLess(np.max(targetindices), ntargets)
+        self.assertLess(np.max(gfaid), 10)
+        self.assertGreaterEqual(np.min(gfaid), 0)
+
+        #- Test higher level get_gfa_targets
+        targets['FLUX_R'] = 2000
+        gfatargets = get_gfa_targets(targets, rfluxlim=1000)
+        self.assertGreater(len(gfatargets), 0)
+        self.assertLess(np.max(gfatargets['GFA_LOC']), 10)
+        self.assertGreaterEqual(np.min(gfatargets['GFA_LOC']), 0)
+        self.assertTrue(tile['TILEID'] in gfatargets['TILEID'])
+
+        #- Shift +60 deg and verify that none are found for that tile
+        targets['RA'] = (targets['RA'] + 60) % 360
+        targetindices, gfaid = on_tile_gfa(tile['TILEID'], targets, 120)
+        self.assertEqual(targetindices.size, 0)
+        self.assertEqual(gfaid.size, 0)
 
 
 def test_suite():