+++ py/desimodel/test/test_io.py
@@ -3,7 +3,9 @@
 """Test desimodel.io.
 """
 import os
+import sys
 import uuid
+import tempfile
 import numpy as np
 from astropy.table import Table
 import unittest
@@ -38,27 +40,42 @@ class TestIO(unittest.TestCase):
         global specter_available, desimodel_available
         cls.specter_available = specter_available
         cls.desimodel_available = desimodel_available
-        cls.trimdir = 'test-'+uuid.uuid4().hex
+        cls.tempdir = tempfile.mkdtemp(prefix='testio-')
+        cls.trimdir = os.path.join(cls.tempdir, 'trim')
+        cls.testfile = os.path.join(cls.tempdir, 'test-abc123.fits')
 
     @classmethod
     def tearDownClass(cls):
-        if os.path.exists(cls.trimdir):
+        if os.path.exists(cls.tempdir):
             import shutil
-            shutil.rmtree(cls.trimdir)
+            shutil.rmtree(cls.tempdir)
 
     def setUp(self):
         """Ensure that any desimodel.io caches are clear before running
         any test.
         """
-        io._thru = dict()
-        io._psf = dict()
-        io._params = None
-        io._gfa = None
-        io._fiberpos = None
-        io._tiles = None
+        io.reset_cache()
+        if os.path.exists(self.testfile):
+            os.remove(self.testfile)
 
     def tearDown(self):
         pass
+
+    def test_reset_cache(self):
+        """Test cache reset (two examples at least)
+        """
+        self.assertTrue(io._fiberpos is None)
+        self.assertTrue(isinstance(io._tiles, dict))
+        self.assertEqual(len(io._tiles), 0)
+        fiberpos = io.load_fiberpos()
+        tiles = io.load_tiles()
+        self.assertTrue(io._fiberpos is not None)
+        self.assertTrue(isinstance(io._tiles, dict))
+        self.assertEqual(len(io._tiles), 1)
+        io.reset_cache()
+        self.assertTrue(io._fiberpos is None)
+        self.assertTrue(isinstance(io._tiles, dict))
+        self.assertEqual(len(io._tiles), 0)
 
     @unittest.skipUnless(specter_available, specter_message)
     def test_load_throughput(self):
@@ -80,6 +97,13 @@ class TestIO(unittest.TestCase):
         """
         p = io.load_desiparams()
         self.assertTrue(isinstance(p, dict))
+
+        #- Samity checks on augmented parameters
+        for channel in ['b', 'r', 'z']:
+            x = p['ccd'][channel]
+            self.assertLess(x['wavemin'], x['wavemax'])
+            self.assertLess(x['wavemin'], x['wavemin_all'])
+            self.assertGreater(x['wavemax'], x['wavemax_all'])
     
     @unittest.skipUnless(desimodel_available, desimodel_message)
     def test_load_gfa(self):
@@ -135,12 +159,18 @@ class TestIO(unittest.TestCase):
     def test_load_tiles(self):
         """Test loading of tile files.
         """
-        self.assertIsNone(io._tiles)
+        # starting clean
+        self.assertEqual(io._tiles, {})
+        t0 = io.load_tiles(cache=False)
+        self.assertEqual(len(io._tiles), 0)
+        # loading tiles fills the cache with one items
         t1 = io.load_tiles(onlydesi=False)
-        tile_cache_id1 = id(io._tiles)
-        self.assertIsNotNone(io._tiles)
+        self.assertEqual(len(io._tiles), 1)
+        tile_cache_id1 = id(list(io._tiles.items())[0])
+        # reloading, even with a filter, shouldn't change cache
         t2 = io.load_tiles(onlydesi=True)
-        tile_cache_id2 = id(io._tiles)
+        self.assertEqual(len(io._tiles), 1)
+        tile_cache_id2 = id(list(io._tiles.items())[0])
         self.assertEqual(tile_cache_id1, tile_cache_id2)
         #- Temporarily support OBSCONDITIONS as u2 (old) or i4 (new)
         self.assertTrue(np.issubdtype(t1['OBSCONDITIONS'].dtype, 'i4') or \
@@ -153,7 +183,7 @@ class TestIO(unittest.TestCase):
         # I think this is the exact same test as above, except using set theory.
         self.assertEqual(len(set(t2.TILEID) - set(t1.TILEID)), 0)
         t3 = io.load_tiles(onlydesi=False)
-        tile_cache_id3 = id(io._tiles)
+        tile_cache_id3 = id(list(io._tiles.items())[0])
         self.assertEqual(tile_cache_id1, tile_cache_id3)
         self.assertTrue(np.issubdtype(t3['OBSCONDITIONS'].dtype, 'i4') or \
                         np.issubdtype(t3['OBSCONDITIONS'].dtype, 'u2') )
@@ -163,6 +193,27 @@ class TestIO(unittest.TestCase):
         b = io.load_tiles(extra=True)
         self.assertGreater(np.sum(np.char.startswith(b['PROGRAM'], 'EXTRA')), 0)
         self.assertLess(len(a), len(b))
+
+    @unittest.skipUnless(desimodel_available, desimodel_message)
+    def test_load_tiles_alt(self):
+        # starting clean
+        self.assertEqual(io._tiles, {})
+        t1 = Table(io.load_tiles())
+        t1.write(self.testfile)
+        #- no path; should fail since that file isn't in $DESIMODEL/data/footprint/
+        if sys.version_info.major == 2:
+            with self.assertRaises(IOError):
+                t2 = io.load_tiles(tilesfile=os.path.basename(self.testfile))
+        else:
+            with self.assertRaises(FileNotFoundError):
+                t2 = io.load_tiles(tilesfile=os.path.basename(self.testfile))
+
+        #- with path, should work:
+        t2 = Table(io.load_tiles(tilesfile=self.testfile))
+        self.assertTrue(np.all(t1 == t2))
+
+        #- cache should have two items now
+        self.assertEqual(len(io._tiles), 2)
 
     @unittest.skipUnless(desimodel_available, desimodel_message)
     def test_tiles_consistency(self):