+++ doc/release_notes.rst
@@ -2,10 +2,32 @@
 desimodel Release Notes
 =======================
 
-0.8.1 (unreleased)
+0.9.2 (unreleased)
 ------------------
 
 * No changes yet.
+
+0.9.1 (2017-11-10)
+------------------
+
+* Extracts wavelength coverage from specpsf files into params dictionary (`PR#68`_).
+* Added desimodel.footprint.program2pass and pass2program to convert between
+  tiling integer pass number and string program name (`PR#67`_).
+
+.. _`PR#67`: https://github.com/desihub/desimodel/pull/67
+.. _`PR#68`: https://github.com/desihub/desimodel/pull/68
+
+0.9.0 (2017-09-19)
+------------------
+
+* Added desimodel.focalplane.radec2xy, which converts RA, Dec coordinates to x, y coordinates on the focal plane, which accepts vector inputs.
+* Added desimodel.focalplane.on_gfa() and its respective helper functions to check if a target is on a GFA of arbitrary telescope pointing
+* Added desimodel.focalplane.on_tile_gfa() to check return a list of indices of targets on a specific tile
+* Added desimodel.focalplane.get_gfa_targets() to return a table with added columns GFA_LOC and TILEID that consists of all targets on any GFA on any tile satisfying a minimum flux in the r-band. 
+* Unittests for the desimodel.focalplane functions were updated accordingly.
+* Added desimodel.footprint.find_points_in_tel_range() to return a list of indices withnin a radius of an arbitray telescope pointing, unaware of tiles (Added respective unittest)
+* Adds desimodel.focalplane.fiber_area_arcsec2()
+* Updates tests to work with trimmed data subset
 
 0.8.0 (2017-08-07)
 ------------------